(MAINT) Prototype Initialize with PDK go template
Prior to this commit all of the logic which depended on the PDK
used the legacy PDK. The Initialize-PuppetModule function, which
called pdk new module, used to take ~30s just to lay down the
files and included numerous unneeded files.

This commit takes advantage of the new work in the PDK go project
and experimentally swaps Initialize-PuppetModule to use it to
scaffold out the new module. This new module lays down a much
smaller number of files - only those which are necessary for the
Puppetization process (including docs generation) and for local
manual testing via bundle execs.

It *also* runs in less than a second. When we move towards the
autogeneration milestone, this will save roughly **three hours**
of build time during a full gallery build/rebuild.

This prototype shows how much faster and slimmer a purpose-built
module template with the new PDK can be.