(MAINT) Define nested cim instance types as enums
Prior to this commit the builder defined the Puppet type for a nested
cim instance's type as a string of its value, ie `'SomeInstanceType'`.

This caused runtime errors when passing the `cim_instance_type` during a
Puppet run as this is not a valid data type.

This commit updates the mapping of a nested CIM Instance into a Puppet
property by defining the `cim_instance_type` as a single-value enum,
such as `Enum['SomeINstanceType']` - this obviates the runtime error.