(GH-186) Handle missing project uri in metadata
Prior to this commit, valid PowerShell modules on the PowerShell Gallery
were sometimes un-Puppetizable as they were missing the `ProjectUri`
metadata key in their manifest.

This commit modifies the logic in the `Get-ReadmeContent` private
function to reuse the module's PowerShell Gallery URI if the Project URI
is null or empty, ensuring that modules with this missing metadata are
Puppetizable.